// src/types/api.ts
export interface paths {
    "/api/rest/issues/{issue_id}": {
        parameters: {
            query?: never;
            header?: never;
            path: {
                /**
                 * @description The issue id
                 * @example {{issueId}}
                 */
                issue_id: string;
            };
            cookie?: never;
        };
        /** Get an issue */
        get: {
// ...
            responses: {
                /** @description Success */
                200: {
// ...
                    content: {
// ...
                        "application/json": {
                            issues?: {
                                id?: number;
                                summary?: string;
                                description?: string;
                                project?: {
                                    id?: number;
                                    name?: string;
                                };
                                category?: {
                                    id?: number;
                                    name?: string;
                                };
                                reporter?: {
                                    id?: number;
                                    name?: string;
                                    real_name?: string;
                                    /** Format: email */
                                    email?: string;
                                };
                                status?: {
                                    id?: number;
                                    name?: string;
                                    label?: string;
                                    /** Format: color */
                                    color?: string;
                                };
                                resolution?: {
                                    id?: number;
                                    name?: string;
                                    label?: string;
                                };
                                view_state?: {
                                    id?: number;
                                    name?: string;
                                    label?: string;
                                };
                                priority?: {
                                    id?: number;
                                    name?: string;
                                    label?: string;
                                };
                                severity?: {
                                    id?: number;
                                    name?: string;
                                    label?: string;
                                };
                                reproducibility?: {
                                    id?: number;
                                    name?: string;
                                    label?: string;
                                };
                                sticky?: boolean;
                                /** Format: date-time */
                                created_at?: string;
                                /** Format: date-time */
                                updated_at?: string;
                                custom_fields?: {
                                    field?: {
                                        id?: number;
                                        name?: string;
                                    };
                                    value?: string;
                                }[];
                                history?: {
                                    /** Format: date-time */
                                    created_at?: string;
                                    user?: {
                                        id?: number;
                                        name?: string;
                                        real_name?: string;
                                        /** Format: email */
                                        email?: string;
                                    };
                                    type?: {
                                        id?: number;
                                        name?: string;
                                    };
                                    message?: string;
                                }[];
                            }[];
                        };
                    };
                };
// ...
            };
        };
// ...
    };
// ...
}